Fix tower hitbox location and fixed fire_tower attack graphic
Tower hitbox was not cospacial with tower image because it was based off
a buggy version of draw for fire tower that had not previously been fixed.
I fixed both.
Fire attack graphic was not disappearing after tower was sold.
Refactored attack method to be more similar to the general use and in
doing so also fixed the graphical issue.