Add gameclock, drawing features, entity, tests
Created and added gameclock to allow decoupling from the pygame
framework. Game control in the main run loop will now be handled by a
gameclock instance.
In level, added in the drawing for the village and background and
changed implementation so that all classes pull from a default village.
Changes to village and background will be put in on a level basis.
Added the entity class and modified the tower and fire tower classes to
be subclasses of entity class.
Added tests for the gameclock and the entity target function.